+++ core/src/components/model/stories/bn-picture-model-viewer.stories.ts
@@ -9,7 +9,22 @@ export default {
 export const Base_Usage = bnPictureModelViewerKit.storyFactory(
   () =>
     // push some child element
-    html`<span>picture-model-viewer demo</span>`,
+    html``,
+  {
+    src: "https://localhost:6006/assets/logo.webp",
+    skyboxImage: "https://localhost:6006/assets/skybox.jpg",
+  },
+).addStyle(`
+bn-picture-model-viewer {
+  width: 80vmin;
+  height: 80vmin;
+}
+`);
+
+export const Custom_Poster = bnPictureModelViewerKit.storyFactory(
+  () =>
+    // push some child element
+    html`<span slot="poster">picture-model-viewer demo</span>`,
   {
     src: "https://localhost:6006/assets/logo.webp",
     skyboxImage: "https://localhost:6006/assets/skybox.jpg",